google-cal/requirements.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0
griptape>=1.1.1
schema>=0.7.5 